Refactored Signup component into multiple compontents for better readability
